+++ src/viser/_scene_api.py
@@ -1097,7 +1097,7 @@ class SceneApi:
         message = _messages.PointCloudMessage(
             name=name,
             props=_messages.PointCloudProps(
-                points=points.astype(np.float16),
+                points=points.astype(np.float32),
                 colors=colors_cast,
                 point_size=point_size,
                 point_ball_norm={
